requirements.txt
ollama>=0.1.0
pyyaml>=6.0
requests>=2.31.0
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.33.0

